+++ src/models/__init__.py
@@ -1,8 +1,11 @@
 
 from .ireview import IReview
 from .gamereview import GameReview
+from .gamereviewembeddings import GameReviewEmbeddings
 
 __all__ = [
     "IReview",
-    "GameReview",
+    "GameReview,
+    GameReviewEmbeddings
+
 ]