+++ CHANGELOG.md
@@ -9,5 +9,6 @@ All notable changes to this project will be documented in this file.
 - Created `SQLiteDatabase` class in `src/data/sqllitedatabase.py`. [dev/aingeru]
 - Created `IReview` class in `src/models/ireview.py`. [dev/aingeru]
 - Created `GameReview` class in `src/models/gamereview.py`. [dev/aingeru]
+- Created `GameReviewEmbeddings` class in `src/models/gamereviewembeddings.py`. [dev/aingeru]
 
 ### Changed